feat: add collector-aware as2rel output alongside classic aggregate
Rewrite as2rel-index to produce BOTH outputs from a single file pass:
1. Classic: {prefix}-latest.json.bz2 (unchanged v1 format — no breaking change)
2. Collector: {prefix}-collector-latest.json.bz2 (new per-collector provenance)

Uses flat vector + sort approach for memory efficiency (~1.7GB peak RSS
for 21M records). All three prefixes (as2rel, as2rel-v4, as2rel-v6) get
both output files.

Remove standalone as2rel-collector-index and pfx2as-collector-index
binaries — collector output is now integrated into as2rel-index.
